Fixed ping issue
