+++ main.go
@@ -8,10 +8,8 @@ import (
 	"os"
 	"os/exec"
 	"os/signal"
-	"strconv"
 	"strings"
 	"syscall"
-	"time"
 
 	"github.com/bwmarrin/discordgo"
 	"github.com/chzyer/readline"
@@ -318,178 +316,6 @@ func printMessage(session *discordgo.Session, msg *discordgo.Message, prefixR bo
 	color.Unset()
 	color.Yellow(s)
 	ColorDefault.Set()
-}
-
-func messageCreate(session *discordgo.Session, e *discordgo.MessageCreate) {
-	channel, err := session.Channel(e.ChannelID)
-	if err != nil {
-		stdutil.PrintErr("Could not get channel", err)
-		return
-	}
-
-	var guild *discordgo.Guild
-	if !channel.IsPrivate {
-		guild, err = session.Guild(channel.GuildID)
-		if err != nil {
-			stdutil.PrintErr("Could not get guild", err)
-			return
-		}
-	}
-
-	if messageCommand(session, e.Message, guild, channel) {
-		return
-	}
-
-	lastMsg = location{
-		guild:   guild,
-		channel: channel,
-	}
-
-	hasOutput := false
-
-	print := false
-Switch:
-	switch messages {
-	case MessagesAll:
-		print = true
-	case MessagesPrivate:
-		if channel.IsPrivate {
-			print = true
-		}
-	case MessagesMentions:
-		if channel.IsPrivate || e.MentionEveryone {
-			print = true
-			break
-		}
-
-		for _, u := range e.Mentions {
-			if u.ID == UserId {
-				print = true
-				break Switch
-			}
-		}
-
-		user, err := session.GuildMember(guild.ID, UserId)
-		if err != nil {
-			stdutil.PrintErr("Could not query user", err)
-			break
-		}
-
-		for _, role := range user.Roles {
-			for _, role2 := range e.MentionRoles {
-				if role == role2 {
-					print = true
-					break Switch
-				}
-			}
-		}
-	case MessagesCurrent:
-		if (guild == nil || loc.guild == nil) && loc.channel != nil && channel.ID != loc.channel.ID {
-			break
-		}
-		if guild != nil && loc.guild != nil && guild.ID != loc.guild.ID {
-			break
-		}
-
-		print = true
-	}
-	if print {
-		printMessage(session, e.Message, true, guild, channel)
-		hasOutput = true
-	}
-
-	if len(luaMessageEvents) > 0 {
-		hasOutput = true
-
-		color.Unset()
-		ColorAutomated.Set()
-
-		fmt.Print("\r" + strings.Repeat(" ", 20) + "\r")
-		luaMessageEvent(session, e.Message)
-
-		color.Unset()
-		ColorDefault.Set()
-	}
-	if hasOutput {
-		printPointer(session)
-	}
-}
-
-func messageCommand(session *discordgo.Session, e *discordgo.Message, guild *discordgo.Guild, channel *discordgo.Channel) (isCmd bool) {
-	if e.Author.ID != UserId {
-		return
-	} else if !intercept {
-		return
-	}
-
-	contents := strings.TrimSpace(e.Content)
-	if !strings.HasPrefix(contents, "console.") {
-		return
-	}
-	cmd := contents[len("console."):]
-
-	isCmd = true
-
-	if strings.EqualFold(cmd, "ping") {
-		first := time.Now().UTC()
-
-		timestamp, err := e.Timestamp.Parse()
-		if err != nil {
-			stdutil.PrintErr("Couldn't parse timestamp", err)
-			return
-		}
-		inMS := first.Sub(timestamp).Nanoseconds() / time.Millisecond.Nanoseconds()
-		text := "Incoming: `" + strconv.FormatInt(inMS, 10) + "ms`"
-
-		middle := time.Now().UTC()
-		_, err = session.ChannelMessageEditComplex(e.ChannelID, e.ID, &discordgo.MessageEdit{
-			Content: "Pong! 1/2",
-			Embed: &discordgo.MessageEmbed{
-				Description: text,
-			},
-		})
-
-		last := time.Now().UTC()
-		outMS := last.Sub(middle).Nanoseconds() / time.Millisecond.Nanoseconds()
-		text += "\nOutgoing: `" + strconv.FormatInt(outMS, 10) + "ms`"
-		text += "\n\n\nIncoming is the time it takes for the message to reach DiscordConsole."
-		text += "\nOutgoing is the time it takes for DiscordConsole to reach discord."
-
-		_, err = session.ChannelMessageEditComplex(e.ChannelID, e.ID, &discordgo.MessageEdit{
-			Content: "Pong! 2/2",
-			Embed: &discordgo.MessageEmbed{
-				Description: text,
-			},
-		})
-		if err != nil {
-			stdutil.PrintErr("Couldn't edit message", err)
-		}
-		return
-	}
-
-	err := session.ChannelMessageDelete(e.ChannelID, e.ID)
-	if err != nil {
-		stdutil.PrintErr("Could not delete message", err)
-	}
-
-	lastLoc = loc
-	loc = location{
-		guild:   guild,
-		channel: channel,
-	}
-	pointerCache = ""
-
-	color.Unset()
-	ColorAutomated.Set()
-
-	fmt.Println(cmd)
-	command(session, cmd)
-
-	color.Unset()
-	ColorDefault.Set()
-
-	printPointer(session)
-	return
 }
 
 const EMPTY_POINTER = "> "
